refactor(cli): extract workflow optimize handler
